Add test cases for test_registrationgui.py
Add a check on register() in registergui.py to flush
any gtk events (and more importantly, wait for any
thing pending in the idle loop). Otherwise there
is a race waiting for the idle timer to check the
threads queue. This was needed to get the test cases
to populate the screens after the threads run.