976855: populate a "version.py" at build time
Instead of asking the rpm db for the versions
of "subsciption-manager" and "python-rhsm", we
generate a versions.py at build time with that
info.

"make set-versions" is the make target that
does it. It will use VERSION from the env
(the spec file passes this in), or if
not specified, it will build one based
on 'git describe'.

This removes code that was on average spending
about .5s every time we started up (including
cli). This reduces it to basically nothing.

This requires latest python-rhsm (>= 1.8.14-1)